pyproject.toml
[project]
name = "gen-ai-hands-on"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-langchain>=0.8.0",
    "ddgs>=9.6.0",
    "docx2txt>=0.9",
    "hf-xet>=1.1.10",
    "langchain-chroma>=0.2.6",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.8",
    "langchain[google-genai]>=0.3.27",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "yfinance>=0.2.66",
]
